+++ adapters/yap.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
 %  Adapter file for YAP Prolog 6.3.4 and later versions
-%  Last updated on November 17, 2023
+%  Last updated on December 2, 2023
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  SPDX-FileCopyrightText: 1998-2023 Paulo Moura <pmoura@logtalk.org>
@@ -224,6 +224,16 @@
 
 '$lgt_prolog_to_logtalk_meta_argument_specifier_hook'(_, _) :-
 	fail.
+
+
+% '$lgt_prolog_phrase_predicate'(@callable)
+%
+% table of predicates that call non-terminals
+% (other than the de facto standard phrase/2-3 predicates)
+
+'$lgt_prolog_phrase_predicate'(pure_input:phrase_from_file(_, _)).
+'$lgt_prolog_phrase_predicate'(pure_input:phrase_from_file(_, _, _)).
+'$lgt_prolog_phrase_predicate'(pure_input:phrase_from_stream(_, _)).
 
 
 % '$lgt_candidate_tautology_or_falsehood_goal_hook'(@callable)